Add makecset logic
This was an attempt to use the split-up technique. Except for very
underconstrained problems where walksat is better anyway it ends up
slower because the csat algorithm itself is too slow.

Idea; Use a better data structure to show the strength of each
connection so we don't have to recompute it.

I think I've done about one lab's worth of work on this now. More still
coming. Probably the rest of this week/next week on SAT optimizations
and then I move to world modeling.